Release 0.16.1
This is a minor version release bringing the fix for the filesystem back-end where it could
not rename a folder (#313).

We also upgrade the proxy-protocol crate that was outstanding as we were waiting for the
maintainer of that project to merge our PR for the upgrade to the bytes 1.0 crate.

Lastly some other dependencies like tokio and hyper were also upgraded.