Allow PROT and PBSZ without requiring auth
This fixes an issue logged in PR #327 where the PHP FTP client issues the
PROT and PBSZ commands to switch the data channel to FTPS but does it
before authenticating. The current libunftp behaviour is to prevent PROT
and PBSZ for unauthenticated clients. This PR then allows it. We might
want to make it a setting though in a PR after this?

Closes #327