Make user non-optional in storage back-ends
This is a breaking change to the StorageBackend trait. I realized while
implementing more storage back-ends in unFTP that there is no need to
make the user detail optional in the trait methods. For example in

```
async fn list<P: AsRef<Path> + Send + Debug>(&self, user: &User, path: P) -> ...
```

the user was previously optional. By the time these methods are called
we will always have a valid user. Its more that the Session struct
stores it as optional and we used it as is. A future refactoring to the
Session struct might split it in two, one for a pre-auth session and one
for after the user was authenticated. There are some nice ways in Rust
to represent different states like this. For now I just fix the storage
back-end to make it more ergonomic and correct.