Support directory timestamps in GCS (#415)
* Support directory timestamps in GCS

Fixes #411

* also support directories that don't have prefix objects

This solves an issue when 'files' were created in some
'subdir' (prefix) without creating a corresponding 'prefix object'.

When you create a directory - through FTP (MKD), or via the Google
Console / Cloud Storage - it creates an object with the name of that
prefix. For instance 'subdir/' will exist as an actual object, with
metadata like any regular object file. The nice thing about that is
that we can use that metadata, to tell when the directory was created.

But in case the file was created through the API, then you can create
a file 'subdir/file', without ever creating a 'subdir/' "prefix
object". This is also reproducible in FTP, by uploading a file to any
non-existent directory. I.e.: there is no problem to 'STOR
/non/existent/dir/file' with unFTP and the GCS backend!

But ofcourse, if we don't have a 'prefix object', then we have no
metadata for that directory. So the code in this commit adds support
for this. Namely, even though there's no object (in items[]) for such
a sub directory, the prefixes[] field _does_ contain it. If there's a
subidr in prefixes[] that isn't in items[], then we will create a fake
Fileinfo entry.