Use a purer Builder pattern for Server
Split Server into Server and ServerBuilder.  The latter has a build()
method that creates the former.  This gives the user the ability to do
stuff after the Server is fully constructed (including the TLS stuff)
but before entering the Server's main loop.  For example, enter
capability mode.

This change also makes the TLS setup apply to Server::service.