fix listing when root path is set
When listing '/' we do not append the `prefix=` filter to the query.
Resulting in a listing of the whole bucket, not the folder we're
confined to.

This fixes that by looking at the 'real' path, i.e. root + request path.
If that is not the root we add the `prefix` filter.

This is quite some effort but having `prefix=/` as filter filters out
all files. So `ls /` wouldn't return anything, but `ls somedir` would
then suddenly work.