Implement 550 for RNFR command
Some FTP clients expect the server to indicate whether a file exists
when they send the RNFR command. If it does, they expect a 350 (which
currently happens). If it doesn't, the expect a 550 ("Requested action
not taken") to indicate the file isn't available.

This adds this behaviour to the RNFR command and adds tests for both
behaviours. I moved some functions from pass_security.rs to common.rs
where I re-use them in the new tests.