Make prometheus and proxy-protocol dependencies optional (#547)
These two dependencies pull in a lot of extra stuff and probably aren't
needed in most common configurations

To make this meaningful we also have to add default-features = false to
the dependencies on libunftp from the other crates. This is technically
a breaking change because e.g. when depending on unftp-sbe-fs but not on
libunftp then none of the two TLS flags may be selected (which doesn't
compile). this seems unlikely to happen in practice.